Move nova connection code to shared library
